+++ src/core/leoCommands.ts
@@ -1979,14 +1979,8 @@ export class Commands {
             }
         }
 
-        let defaultAbsbase = g.app.homeDir || process.cwd();
-        if (g.isBrowser) {
-            // In the browser, we use the base of the repo with a single slash "/"
-            defaultAbsbase = "/";
-        }
-
         // Add absbase and reverse the list.
-        const absbase = c.fileName() ? g.os_path_dirname(c.fileName()) : defaultAbsbase;
+        const absbase = c.fileName() ? g.os_path_dirname(c.fileName()) : g.app.homeDir!;
         paths.push(absbase);
         paths.reverse();